Workaround against Java 8 vs 9 binary incompatibility
In Java >= 9 position() method has been overridden in ByteBuffer and made
returning covariant ByteBuffer type. This causes the new compiler to generate
a call using the newer signature. In its turn the compiled code stops working
on older JVM.

We're supporting OpenHAB v2, which uses JRE 1.8; and we don't have control over
what's in PATH when compiling from the command line. The only option is to do
this ugly workaround. Thank you Oracle!